Attempted to simplify lsp loading - also made lazy
