+++ dot_config/nvim/lua/tap/plugins/lsp/servers/bashls.lua
@@ -1,9 +1,11 @@
 local lsp_utils = require 'tap.utils.lsp'
 
-local module = {}
+local M = {}
 
-function module.setup()
+M.ensure_installed = { 'bashls' }
+
+function M.setup()
   require('lspconfig').bashls.setup(lsp_utils.merge_with_default_config())
 end
 
-return module
+return M
