+++ dot_config/nvim/lua/tap/plugins/lsp/servers/init.lua
@@ -1,0 +1,8 @@
+return {
+  require 'tap.plugins.lsp.servers.bashls',
+  require 'tap.plugins.lsp.servers.eslint',
+  require 'tap.plugins.lsp.servers.jsonls',
+  require 'tap.plugins.lsp.servers.rnix',
+  require 'tap.plugins.lsp.servers.sumneko_lua',
+  require 'tap.plugins.lsp.servers.tsserver',
+}
